# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QRX API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sliders:
    get:
      operationId: listSliders
      tags: [sliders]
      summary: List active sliders (public); pass all=true for admin
      parameters:
        - name: position
          in: query
          schema:
            type: string
        - name: all
          in: query
          description: Admin only — include inactive sliders
          schema:
            type: string
      responses:
        "200":
          description: Slider list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slider"
    post:
      operationId: createSlider
      tags: [sliders]
      summary: Create slider (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SliderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Slider"
  /sliders/{id}:
    patch:
      operationId: updateSlider
      tags: [sliders]
      summary: Update slider (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SliderInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Slider"
    delete:
      operationId: deleteSlider
      tags: [sliders]
      summary: Delete slider (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- VIDEO PROMOTIONS ----
...
components:
  schemas:
...
    Slider:
      type: object
      required: [id, title, position, isActive, autoPlay, slideInterval, showArrows, showDots, priority, displayOrder, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        buttonText:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        position:
          type: string
        isActive:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        priority:
          type: integer
        displayOrder:
          type: integer
        autoPlay:
          type: boolean
        slideInterval:
          type: integer
        showArrows:
          type: boolean
        showDots:
          type: boolean
        desktopWidth:
          type: ["integer", "null"]
        desktopHeight:
          type: ["integer", "null"]
        mobileWidth:
          type: ["integer", "null"]
        mobileHeight:
          type: ["integer", "null"]
        tabletWidth:
          type: ["integer", "null"]
        tabletHeight:
          type: ["integer", "null"]
        customWidth:
          type: ["integer", "null"]
        customHeight:
          type: ["integer", "null"]
        createdAt:
          type: string
    SliderInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        buttonText:
          type: string
        description:
          type: string
        position:
          type: string
        isActive:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        priority:
          type: integer
        displayOrder:
          type: integer
        autoPlay:
          type: boolean
        slideInterval:
          type: integer
        showArrows:
          type: boolean
        showDots:
          type: boolean
        desktopWidth:
          type: ["integer", "null"]
        desktopHeight:
          type: ["integer", "null"]
        mobileWidth:
          type: ["integer", "null"]
        mobileHeight:
          type: ["integer", "null"]
        tabletWidth:
          type: ["integer", "null"]
        tabletHeight:
          type: ["integer", "null"]
        customWidth:
          type: ["integer", "null"]
        customHeight:
          type: ["integer", "null"]
...